refactor: remove ProtectedComponent wrappers from various components
