+++ src/app/dashboard/admin/quotation-groups/_components/QuotationGroupPrimaryButtons.tsx
@@ -2,10 +2,8 @@
 
 import { Plus } from "lucide-react";
 
-import { ProtectedComponent } from "@/auth/presentation/components/ProtectedComponent";
 import { Button } from "@/shared/components/ui/button";
 import { useDialogStore } from "@/shared/stores/useDialogStore";
-import { EnumAction, EnumResource } from "../../roles/_utils/groupedPermission";
 import { MODULE_QUOTATION_GROUP } from "./QuotationGroupDialogs";
 
 export default function QuotationGroupPrimaryButtons() {
@@ -15,7 +13,7 @@ export default function QuotationGroupPrimaryButtons() {
 
   return (
     <div className="flex gap-2">
-      <ProtectedComponent requiredPermissions={[{ resource: EnumResource.quotations, action: EnumAction.create }]}>
+      <>
         <Button
           className="space-x-1"
           onClick={() => {
@@ -25,7 +23,7 @@ export default function QuotationGroupPrimaryButtons() {
           Crear grupo de cotizaciones
           <Plus className="h-4 w-4" />
         </Button>
-      </ProtectedComponent>
+      </>
     </div>
   );
 }
