+++ src/app/dashboard/admin/quotation/_components/QuotationPrimaryButtons.tsx
@@ -2,10 +2,8 @@
 
 import { Plus } from "lucide-react";
 
-import { ProtectedComponent } from "@/auth/presentation/components/ProtectedComponent";
 import { Button } from "@/shared/components/ui/button";
 import { useDialogStore } from "@/shared/stores/useDialogStore";
-import { EnumAction, EnumResource } from "../../roles/_utils/groupedPermission";
 
 export default function QuotationPrimaryButtons() {
   const { open } = useDialogStore();
@@ -15,7 +13,7 @@ export default function QuotationPrimaryButtons() {
 
   return (
     <div>
-      <ProtectedComponent requiredPermissions={[{ resource: EnumResource.quotations, action: EnumAction.create }]}>
+      <>
         <Button
           className="space-x-1"
           onClick={() => {
@@ -24,7 +22,7 @@ export default function QuotationPrimaryButtons() {
         >
           <span>Agregar Cotización</span> <Plus className="h-4 w-4" />
         </Button>
-      </ProtectedComponent>
+      </>
     </div>
   );
 }
