+++ src/app/dashboard/admin/quotation/_components/QuotationTableActions.tsx
@@ -2,7 +2,6 @@
 
 import { Edit, MoreHorizontal, Trash } from "lucide-react";
 
-import { ProtectedComponent } from "@/auth/presentation/components/ProtectedComponent";
 import { Button } from "@/shared/components/ui/button";
 import {
   DropdownMenu,
@@ -13,7 +12,6 @@ import {
 } from "@/shared/components/ui/dropdown-menu";
 import { useDialogStore } from "@/shared/stores/useDialogStore";
 import { QuotationResponse } from "../_types/quotation.types";
-import { EnumAction, EnumResource } from "../../roles/_utils/groupedPermission";
 
 interface QuotationTableActionsProps {
   quotation: QuotationResponse;
@@ -34,12 +32,7 @@ export default function QuotationTableActions({ quotation }: QuotationTableActio
   };
 
   return (
-    <ProtectedComponent
-      requiredPermissions={[
-        { resource: EnumResource.quotations, action: EnumAction.update },
-        { resource: EnumResource.quotations, action: EnumAction.delete },
-      ]}
-    >
+    <>
       <div className="flex items-center gap-2">
         <DropdownMenu>
           <DropdownMenuTrigger asChild>
@@ -48,29 +41,25 @@ export default function QuotationTableActions({ quotation }: QuotationTableActio
             </Button>
           </DropdownMenuTrigger>
           <DropdownMenuContent>
-            <ProtectedComponent
-              requiredPermissions={[{ resource: EnumResource.quotations, action: EnumAction.update }]}
-            >
+            <>
               <DropdownMenuItem className="cursor-pointer" onClick={handleEdit}>
                 Editar
                 <DropdownMenuShortcut>
                   <Edit className="size-4 mr-2" />
                 </DropdownMenuShortcut>
               </DropdownMenuItem>
-            </ProtectedComponent>
-            <ProtectedComponent
-              requiredPermissions={[{ resource: EnumResource.quotations, action: EnumAction.delete }]}
-            >
+            </>
+            <>
               <DropdownMenuItem className="cursor-pointer" onClick={handleDelete} disabled={quotation.isConcrete}>
                 Eliminar
                 <DropdownMenuShortcut>
                   <Trash className="size-4 mr-2" />
                 </DropdownMenuShortcut>
               </DropdownMenuItem>
-            </ProtectedComponent>
+            </>
           </DropdownMenuContent>
         </DropdownMenu>
       </div>
-    </ProtectedComponent>
+    </>
   );
 }
