+++ src/app/dashboard/admin/roles/_components/RolesPrimaryButtons.tsx
@@ -2,10 +2,8 @@
 
 import { Plus } from "lucide-react";
 
-import { ProtectedComponent } from "@/auth/presentation/components/ProtectedComponent";
 import { Button } from "@/shared/components/ui/button";
 import { useDialogStore } from "@/shared/stores/useDialogStore";
-import { EnumAction, EnumResource } from "../_utils/groupedPermission";
 
 export default function RolesPrimaryButtons() {
   const { open } = useDialogStore();
@@ -14,7 +12,7 @@ export default function RolesPrimaryButtons() {
   const MODULE = "roles";
 
   return (
-    <ProtectedComponent requiredPermissions={[{ resource: EnumResource.roles, action: EnumAction.create }]}>
+    <>
       <div className="flex gap-2">
         <Button
           className="space-x-1"
@@ -25,6 +23,6 @@ export default function RolesPrimaryButtons() {
           <span>Agregar Rol</span> <Plus size={18} />
         </Button>
       </div>
-    </ProtectedComponent>
+    </>
   );
 }
