+++ src/app/dashboard/admin/roles/_components/RolesTableActions.tsx
@@ -2,7 +2,6 @@
 
 import { Edit, MoreHorizontal, RotateCcw, Trash } from "lucide-react";
 
-import { ProtectedComponent } from "@/auth/presentation/components/ProtectedComponent";
 import { Button } from "@/shared/components/ui/button";
 import {
   DropdownMenu,
@@ -14,7 +13,6 @@ import {
 import { useDialogStore } from "@/shared/stores/useDialogStore";
 import { useToggleActiveRole } from "../_hooks/useRoles";
 import { Role } from "../_types/roles";
-import { EnumAction, EnumResource } from "../_utils/groupedPermission";
 
 interface RolesTableActionsProps {
   row: Role;
@@ -41,12 +39,7 @@ export function RolesTableActions({ row }: RolesTableActionsProps) {
   };
 
   return (
-    <ProtectedComponent
-      requiredPermissions={[
-        { resource: EnumResource.roles, action: EnumAction.update },
-        { resource: EnumResource.roles, action: EnumAction.delete },
-      ]}
-    >
+    <>
       <DropdownMenu>
         <DropdownMenuTrigger asChild>
           <Button variant="ghost" className="h-8 w-8 p-0">
@@ -55,15 +48,15 @@ export function RolesTableActions({ row }: RolesTableActionsProps) {
           </Button>
         </DropdownMenuTrigger>
         <DropdownMenuContent align="end">
-          <ProtectedComponent requiredPermissions={[{ resource: EnumResource.roles, action: EnumAction.update }]}>
+          <>
             <DropdownMenuItem onClick={handleEdit}>
               Editar
               <DropdownMenuShortcut>
                 <Edit className="mr-2 h-4 w-4" />
               </DropdownMenuShortcut>
             </DropdownMenuItem>
-          </ProtectedComponent>
-          <ProtectedComponent requiredPermissions={[{ resource: EnumResource.roles, action: EnumAction.delete }]}>
+          </>
+          <>
             {row.isActive ? (
               <DropdownMenuItem onClick={handleDelete} disabled={!row.isActive}>
                 Eliminar
@@ -79,9 +72,9 @@ export function RolesTableActions({ row }: RolesTableActionsProps) {
                 </DropdownMenuShortcut>
               </DropdownMenuItem>
             )}
-          </ProtectedComponent>
+          </>
         </DropdownMenuContent>
       </DropdownMenu>
-    </ProtectedComponent>
+    </>
   );
 }
