+++ src/app/dashboard/admin/services/_components/ActivitiesList.tsx
@@ -2,8 +2,6 @@ import { useEffect } from "react";
 import { useQueryClient } from "@tanstack/react-query";
 import { ChevronRight, Loader2, Paperclip, Plus, SquareDashed } from "lucide-react";
 
-import { ProtectedComponent } from "@/auth/presentation/components/ProtectedComponent";
-import { useAuth } from "@/auth/presentation/providers/AuthProvider";
 import { Badge } from "@/shared/components/ui/badge";
 import { Button } from "@/shared/components/ui/button";
 import { ScrollArea } from "@/shared/components/ui/scroll-area";
@@ -13,7 +11,6 @@ import { ACTIVITIES_KEYS, useActivitiesByObjectiveId } from "../_hooks/useActivi
 import { OBJECTIVES_KEYS } from "../_hooks/useObjectives";
 import { SERVICES_KEYS, useServices } from "../_hooks/useServices";
 import { useServiceStore } from "../_hooks/useServiceStore";
-import { EnumAction, EnumResource } from "../../roles/_utils/groupedPermission";
 import CardItem from "./CardItem";
 
 export default function ActivitiesList() {
@@ -24,7 +21,6 @@ export default function ActivitiesList() {
   const { data: services } = useServices();
   const { open, isOpenForModule } = useDialogStore();
   const queryClient = useQueryClient();
-  const { hasPermission } = useAuth();
   // Refrescar la lista cuando cambian los servicios (por si el objetivo actual tiene nuevas actividades)
   useEffect(() => {
     if (selectedObjective && services) {
@@ -75,7 +71,7 @@ export default function ActivitiesList() {
         <div className="flex flex-col gap-2">
           <h3 className="text-lg font-bold">Lista de Actividades</h3>
         </div>
-        <ProtectedComponent requiredPermissions={[{ resource: EnumResource.services, action: EnumAction.create }]}>
+        <>
           <Button
             size="sm"
             disabled={!selectedObjective}
@@ -85,7 +81,7 @@ export default function ActivitiesList() {
             <Plus className="w-4 h-4" />
             Nueva Actividad
           </Button>
-        </ProtectedComponent>
+        </>
       </div>
       <Separator />
       {error && <div className="flex items-center justify-center h-full">{error.message}</div>}
@@ -125,10 +121,6 @@ export default function ActivitiesList() {
                         </Badge>
                       )
                     }
-                    permissions={{
-                      update: hasPermission(EnumResource.services, EnumAction.update),
-                      delete: hasPermission(EnumResource.services, EnumAction.delete),
-                    }}
                   />
                 ))
               ) : (
