+++ src/app/dashboard/admin/services/_components/ServicesList.tsx
@@ -1,7 +1,5 @@
 import { Loader2, Plus, SquareDashed } from "lucide-react";
 
-import { ProtectedComponent } from "@/auth/presentation/components/ProtectedComponent";
-import { useAuth } from "@/auth/presentation/providers/AuthProvider";
 import { Badge } from "@/shared/components/ui/badge";
 import { Button } from "@/shared/components/ui/button";
 import { ScrollArea } from "@/shared/components/ui/scroll-area";
@@ -11,14 +9,12 @@ import { useDialogStore } from "@/shared/stores/useDialogStore";
 import { useServices } from "../_hooks/useServices";
 import { useServiceStore } from "../_hooks/useServiceStore";
 import { ServiceResponse } from "../_types/services.types";
-import { EnumAction, EnumResource } from "../../roles/_utils/groupedPermission";
 import CardItem from "./CardItem";
 
 export default function ServicesList() {
   const { data: services, isLoading, error } = useServices();
   const { open } = useDialogStore();
   const { setSelectedService, selectedService, setSelectedObjective, setSelectedActivity } = useServiceStore();
-  const { hasPermission } = useAuth();
   if (error) return <div>Error: {error.message}</div>;
 
   const handleDelete = (service: ServiceResponse) => {
@@ -44,12 +40,12 @@ export default function ServicesList() {
           <h3 className="text-lg font-bold">Lista de Servicios</h3>
           <p className="text-sm text-muted-foreground">Aquí puedes ver todos los servicios que tienes disponibles.</p>
         </div>
-        <ProtectedComponent requiredPermissions={[{ resource: EnumResource.services, action: EnumAction.create }]}>
+        <>
           <Button size="sm" disabled={isLoading} variant="outline" onClick={() => open("services", "create")}>
             <Plus className="w-4 h-4" />
             Nuevo Servicio
           </Button>
-        </ProtectedComponent>
+        </>
       </div>
       <Separator />
       {isLoading ? (
@@ -70,10 +66,6 @@ export default function ServicesList() {
                   onDelete={() => handleDelete(service)}
                   onClick={() => handleServiceClick(service)}
                   className={cn(selectedService?.id === service.id && "border-sky-400  outline-4 outline-sky-300/10")}
-                  permissions={{
-                    update: hasPermission(EnumResource.services, EnumAction.update),
-                    delete: hasPermission(EnumResource.services, EnumAction.delete),
-                  }}
                 />
               ))
             ) : (
