+++ src/app/dashboard/admin/tracking/[serviceId]/_components/ProjectActivitiesActions.tsx
@@ -1,7 +1,5 @@
 import { Edit, MoreHorizontal, Trash } from "lucide-react";
 
-import { EnumAction, EnumResource } from "@/app/dashboard/admin/roles/_utils/groupedPermission";
-import { ProtectedComponent } from "@/auth/presentation/components/ProtectedComponent";
 import { Button } from "@/shared/components/ui/button";
 import {
   DropdownMenu,
@@ -40,12 +38,7 @@ export default function ProjectActivitiesActions({ row }: ProjectActivitiesActio
   };
 
   return (
-    <ProtectedComponent
-      requiredPermissions={[
-        { resource: EnumResource.projects, action: EnumAction.edit },
-        { resource: EnumResource.projects, action: EnumAction.delete },
-      ]}
-    >
+    <>
       <DropdownMenu>
         <DropdownMenuTrigger asChild>
           <Button variant="ghost" className="bg-background" size="icon">
@@ -53,24 +46,24 @@ export default function ProjectActivitiesActions({ row }: ProjectActivitiesActio
           </Button>
         </DropdownMenuTrigger>
         <DropdownMenuContent>
-          <ProtectedComponent requiredPermissions={[{ resource: EnumResource.projects, action: EnumAction.edit }]}>
+          <>
             <DropdownMenuItem onClick={handleEdit}>
               Editar
               <DropdownMenuShortcut>
                 <Edit className="size-4" />
               </DropdownMenuShortcut>
             </DropdownMenuItem>
-          </ProtectedComponent>
-          <ProtectedComponent requiredPermissions={[{ resource: EnumResource.projects, action: EnumAction.delete }]}>
+          </>
+          <>
             <DropdownMenuItem onClick={handleDelete}>
               Eliminar
               <DropdownMenuShortcut>
                 <Trash className="size-4" />
               </DropdownMenuShortcut>
             </DropdownMenuItem>
-          </ProtectedComponent>
+          </>
         </DropdownMenuContent>
       </DropdownMenu>
-    </ProtectedComponent>
+    </>
   );
 }
