+++ src/app/dashboard/admin/tracking/[serviceId]/page.tsx
@@ -5,8 +5,6 @@ import Link from "next/link";
 import { useParams, useRouter } from "next/navigation";
 import { toast } from "sonner";
 
-import { ProtectedComponent } from "@/auth/presentation/components/ProtectedComponent";
-import AlertMessage from "@/shared/components/alerts/Alert";
 import { ShellHeader, ShellTitle } from "@/shared/components/layout/Shell";
 import { LoadingOverlay } from "@/shared/components/loading-overlay";
 import {
@@ -18,7 +16,6 @@ import {
   BreadcrumbSeparator,
 } from "@/shared/components/ui/breadcrumb";
 import { useServiceById } from "../_hooks/useServicesProject";
-import { EnumAction, EnumResource } from "../../roles/_utils/groupedPermission";
 import { ListObjectives } from "./_components/ListObjectives";
 import ProjectObjectivesDialogs from "./_components/ProjectObjectivesDialogs";
 import ProjectObjectivesPrimaryButtons from "./_components/ProjectObjectivesPrimaryButtons";
@@ -41,16 +38,7 @@ export default function AdminServiceDetailPage() {
     return <LoadingOverlay isLoading={isLoading} className="size-full min-h-screen" />;
   }
   return (
-    <ProtectedComponent
-      requiredPermissions={[{ resource: EnumResource.projects, action: EnumAction.read }]}
-      fallback={
-        <AlertMessage
-          variant="destructive"
-          title="No tienes permisos para ver este contenido"
-          description="Por favor, contacta al administrador. O intenta iniciar sesión con otro usuario."
-        />
-      }
-    >
+    <>
       <Breadcrumb className="mb-4">
         <BreadcrumbList>
           <BreadcrumbItem>
@@ -70,6 +58,6 @@ export default function AdminServiceDetailPage() {
       </ShellHeader>
       <ListObjectives serviceId={serviceId as string} />
       <ProjectObjectivesDialogs serviceId={serviceId as string} />
-    </ProtectedComponent>
+    </>
   );
 }
