+++ src/app/dashboard/admin/tracking/_components/ProjectServiceCard.tsx
@@ -1,7 +1,6 @@
 import { useRouter } from "next/navigation";
 import { CheckCircle, Circle, Clock, Edit, MoreVertical, Trash } from "lucide-react";
 
-import { ProtectedComponent } from "@/auth/presentation/components/ProtectedComponent";
 import { Button } from "@/shared/components/ui/button";
 import { Card, CardDescription, CardFooter, CardHeader, CardTitle } from "@/shared/components/ui/card";
 import {
@@ -15,7 +14,6 @@ import { Progress } from "@/shared/components/ui/progress";
 import { cn } from "@/shared/lib/utils";
 import { useDialogStore } from "@/shared/stores/useDialogStore";
 import { ProjectServiceResponse } from "../_types/tracking.types";
-import { EnumAction, EnumResource } from "../../roles/_utils/groupedPermission";
 import { PROJECT_SERVICE_MODULE } from "./ProjectsDialogs";
 
 interface ProjectServiceCardProps {
@@ -57,12 +55,7 @@ export default function ProjectServiceCard({ service }: ProjectServiceCardProps)
               <span className="first-letter:uppercase">{service.name}</span>
             </div>
           </div>
-          <ProtectedComponent
-            requiredPermissions={[
-              { resource: EnumResource.projects, action: EnumAction.edit },
-              { resource: EnumResource.projects, action: EnumAction.delete },
-            ]}
-          >
+          <>
             <DropdownMenu>
               <DropdownMenuTrigger asChild>
                 <Button variant="ghost" size="icon">
@@ -70,9 +63,7 @@ export default function ProjectServiceCard({ service }: ProjectServiceCardProps)
                 </Button>
               </DropdownMenuTrigger>
               <DropdownMenuContent>
-                <ProtectedComponent
-                  requiredPermissions={[{ resource: EnumResource.projects, action: EnumAction.edit }]}
-                >
+                <>
                   <DropdownMenuItem
                     onClick={(e) => {
                       e.stopPropagation();
@@ -84,10 +75,8 @@ export default function ProjectServiceCard({ service }: ProjectServiceCardProps)
                       <Edit />
                     </DropdownMenuShortcut>
                   </DropdownMenuItem>
-                </ProtectedComponent>
-                <ProtectedComponent
-                  requiredPermissions={[{ resource: EnumResource.projects, action: EnumAction.delete }]}
-                >
+                </>
+                <>
                   <DropdownMenuItem
                     onClick={(e) => {
                       e.stopPropagation();
@@ -99,10 +88,10 @@ export default function ProjectServiceCard({ service }: ProjectServiceCardProps)
                       <Trash />
                     </DropdownMenuShortcut>
                   </DropdownMenuItem>
-                </ProtectedComponent>
+                </>
               </DropdownMenuContent>
             </DropdownMenu>
-          </ProtectedComponent>
+          </>
         </CardTitle>
         <CardDescription className="flex flex-col gap-2 ">{service.description}</CardDescription>
       </CardHeader>
